refactor(quest): 一覧ページのリファクタリング
